# pkg/api/docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  rest.CreateSystemRequest:
    properties:
      definitionUrl:
        type: string
      id:
        $ref: '#/definitions/v1.SystemID'
        type: object
    type: object
...
  rest.SetSecretRequest:
    properties:
      value:
        type: string
    type: object
...
  v1.Result:
    type: object
  v1.Secret:
    properties:
      name:
        description: |
          Name
        type: string
      path:
        description: |
          Secret service path
        type: string
      value:
        description: |
          Value
        type: string
    type: object
  v1.Service:
    properties:
      availableInstances:
        description: |
          TBD
        type: integer
      failureInfo:
        $ref: '#/definitions/v1.ServiceFailureInfo'
        description: |
          TBD
        type: object
      id:
        $ref: '#/definitions/v1.ServiceID'
        description: |
          Service ID
        type: object
      instances:
        description: |
          TBD
        items:
          type: string
        type: array
      message:
        description: |
          TBD
        type: string
      path:
        description: |
          Service Path
        type: string
      ports:
        description: |
          TBD
        type: object
      staleInstances:
        description: |
          TBD
        type: integer
      state:
        $ref: '#/definitions/v1.ServiceState'
        description: |
          State ["pending", "deleting", "scaling", "updating", "stable", "failed"]
        type: object
      terminatingInstances:
        description: |
          TBD
        type: integer
      updatedInstances:
        description: |
          TBD
        type: integer
    type: object
  v1.ServiceFailureInfo:
    properties:
      message:
        type: string
      time:
        type: string
    type: object
  v1.ServiceID:
    type: object
  v1.ServiceState:
    type: object
  v1.System:
    properties:
      definitionUrl:
        description: |
          git url for for where the definition lives in
        type: string
      id:
        $ref: '#/definitions/v1.SystemID'
        description: |
          System ID
        type: object
      services:
        description: |
          map for service path and services currently running in the system
        type: object
      state:
        $ref: '#/definitions/v1.SystemState'
        description: |
          State. One of "pending", "failed", "deleting", "stable", "degraded", "scaling", "updating"
        type: object
    type: object
  v1.SystemID:
    type: object
  v1.SystemState:
    type: object
...
host: localhost:8876
info:
  contact: {}
  description: Welcome to lattice api
  license:
    name: Apache 2.0
  termsOfService: TBD
  title: Lattice API
  version: "1.0"
paths:
  /systems:
    get:
      consumes:
      - application/json
      description: list systems
      operationId: list-systems
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.System'
            type: array
      summary: Lists systems
    post:
      consumes:
      - application/json
      description: create system
      operationId: create-system
      parameters:
      - description: Create system
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.CreateSystemRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.System'
            type: object
      summary: Create a new system
  /systems/{id}:
    delete:
      consumes:
      - application/json
      description: get system
      operationId: delete-system
      parameters:
      - description: System ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Result'
            type: object
      summary: Delete system
    get:
      consumes:
      - application/json
      description: get system
      operationId: get-system
      parameters:
      - description: System ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.System'
            type: object
      summary: Get system
...
  /systems/{systemId}/secrets:
    get:
      consumes:
      - application/json
      description: list secrets
      operationId: list-secrets
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Secret'
            type: array
      summary: Lists secrets
    post:
      consumes:
      - application/json
      description: set secret
      operationId: set-secret
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Create secret
        in: body
        name: secretRequest
        required: true
        schema:
          $ref: '#/definitions/rest.SetSecretRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Secret'
            type: object
      summary: set secret
  /systems/{systemId}/secrets/{id}:
    delete:
      consumes:
      - application/json
      description: get system
      operationId: delete-system
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Secret ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Result'
            type: object
      summary: Delete system
    get:
      consumes:
      - application/json
      description: get secret
      operationId: get-secret
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Secret ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Secret'
            type: object
      summary: Get secret
  /systems/{systemId}/services:
    get:
      consumes:
      - application/json
      description: list services
      operationId: list-services
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Service'
            type: array
      summary: Lists services
  /systems/{systemId}/services/{id}:
    get:
      consumes:
      - application/json
      description: get service
      operationId: get-service
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Service'
            type: object
      summary: Get service
  /systems/{systemId}/services/{id}/logs:
    get:
      consumes:
      - application/json
      description: get service logs
      operationId: get-service-logs
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Instance
        in: query
        name: instance
        required: true
        type: string
      - description: Sidecar
        in: query
        name: sidecar
        type: string
      - description: Follow
        in: query
        name: follow
        type: string
      - description: Previous
        in: query
        name: previous
        type: boolean
      - description: Timestamps
        in: query
        name: timestamps
        type: boolean
      - description: tail
        in: query
        name: tail
        type: integer
      - description: Since
        in: query
        name: since
        type: string
      - description: Since Time
        in: query
        name: sinceTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: log stream
          schema:
            type: string
      summary: Get service logs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apiKey
    type: apiKey
swagger: "2.0"